chore(core): rebase
gh-1922
