+++ packages/core/src/component.ts
@@ -38,9 +38,8 @@ import {IncomingMessage, ServerResponse} from 'http';
 
 import {TenantContextMiddlewareInterceptorProvider} from './middlewares';
 import {TenantIdEncryptionProvider} from './providers';
-import {AwsSsmHelperService} from './services';
+import {AwsSsmHelperService, TenantIdentifierHelperService} from './services';
 import {CoreConfig, addTenantId} from './types';
-
 export class CoreComponent implements Component {
   constructor(
     @inject(CoreBindings.APPLICATION_INSTANCE)
@@ -110,6 +109,9 @@ export class CoreComponent implements Component {
     this.bindings.push(Binding.bind(SFCoreBindings.i18n).to(this.localeObj));
     this.application.add(createBindingFromClass(OperationSpecEnhancer));
     this.application.bind('AwsSsmHelperService').toClass(AwsSsmHelperService);
+    this.application
+      .bind('TenantIdentifierHelperService')
+      .toClass(TenantIdentifierHelperService);
 
     this.application.component(Loopback4DynamicDatasourceComponent);
     if (process.env.SAAS_MODEL === 'silo storage') {
