+++ packages/core/src/services/index.ts
@@ -1,1 +1,2 @@
 export * from './aws-ssm-helper.service';
+export * from './tenant-identifier-helper.service';
